requirements.txt
face-recognition>=1.3.0
Pillow>=10.0.0
numpy>=1.24.0
dlib>=19.24.0
setuptools>=65.0.0
git+https://github.com/ageitgey/face_recognition_models

